Add sha3 python package to cron tasks virtualenv
This is needed by the RSA package when running on Python 3.5 (which we
are for now on Debian Stretch). SHA3 is supported natively by Python 3.6
and above, but requires a third-party library with Python 3.5, so this
dependency can be removed once we're fully upgraded.

See: https://github.com/sybrenstuvel/python-rsa/commit/3c5ee5