Dockerfile: Update to Debian Bullseye and Mediawiki 1.35
This updates to Bullseye and thereby to Mediawiki 1.35. It also updates
the extensions we use to their current state of the art.

For now sticking with Composer 1.x as Composer 2.x complains over the
version of `wikimedia/composer-merge-plugin`. The easiest solution for
now seems to be to install Composer 1.x and run `update`.