Update uritemplate for more recent Python support
This bumps uritemplate to 3.0.1, which should have a fix for the
following deprecation warning:

`Using or importing the ABCs from 'collections' instead of from
'collections.abc' is deprecated since Python 3.3, and in 3.10 it will
stop working`.

Can't bump this further just yet due to other package conflicts with
`google-api-python-client` that will need reviewing.